AG-13929 - Add ag-charts-server-side package implementation
