+++ packages/ag-charts-server-side/src/agChartsServerSide.ts
@@ -1,18 +1,13 @@
-import type { ExportFormat } from 'skia-canvas';
-import { FontLibrary, Image } from 'skia-canvas';
+import { FontLibrary } from 'skia-canvas';
 
 import { AgCharts } from 'ag-charts-community';
 
-import { NodeCanvas } from './canvas-config';
+import { NodeCanvas } from './canvasConfig';
+import { patchDocumentCreateElement } from './documentPatch';
 import { createIsolatedEnvironment } from './environment';
 import type { FinancialChartRenderOptions, FontDefinition, GaugeRenderOptions, RenderOptions } from './types';
 
 const DEFAULT_TIMEOUT = 30000;
-
-/** Document with overridable createElement for mock canvas injection */
-interface MockableDocument extends Document {
-    createElement: (tagName: string, options?: ElementCreationOptions) => HTMLElement;
-}
 
 // Module-level mutex to serialize renders (prevents global document/window races)
 let renderLock: Promise<void> = Promise.resolve();
@@ -93,40 +88,9 @@ export class AgChartsServerSide {
         let chart: { destroy(): void } | undefined;
 
         try {
-            const doc = env.document as MockableDocument;
-            const realCreateElement = doc.createElement.bind(doc);
-            doc.createElement = (tag: string, opts?: ElementCreationOptions): HTMLElement => {
-                if (tag === 'canvas') {
-                    const canvas = canvasStack.shift() ?? new NodeCanvas(width * pixelRatio, height * pixelRatio);
-                    const mockElement = realCreateElement(tag, opts) as HTMLCanvasElement;
-
-                    const originalGetContext = mockElement.getContext.bind(mockElement);
-                    Object.defineProperty(mockElement, 'getContext', {
-                        value: (contextId: string, _options?: unknown) => {
-                            if (contextId === '2d') {
-                                return canvas.getContext('2d');
-                            }
-                            return originalGetContext(contextId as '2d');
-                        },
-                        writable: true,
-                        configurable: true,
-                    });
-
-                    Object.defineProperty(mockElement, 'toDataURL', {
-                        value: (mimeType = 'image/png') => {
-                            return canvas.toDataURLSync(mimeType.split('/')[1] as ExportFormat);
-                        },
-                        writable: true,
-                        configurable: true,
-                    });
-
-                    return mockElement;
-                }
-                if (tag === 'img') {
-                    return new Image() as unknown as HTMLImageElement;
-                }
-                return realCreateElement(tag, opts);
-            };
+            patchDocumentCreateElement(env.document, {
+                getCanvas: () => canvasStack.shift() ?? new NodeCanvas(width * pixelRatio, height * pixelRatio),
+            });
 
             const container = env.document.getElementById('container')!;
