+++ packages/ag-charts-server-side/src/canvasConfig.ts
