+++ packages/ag-charts-server-side/src/test/index.ts
@@ -1,3 +1,3 @@
 export { setupMockCanvas, type MockCanvasContext } from './setupMockCanvas';
 export { CANVAS_WIDTH, CANVAS_HEIGHT, CANVAS_TO_BUFFER_DEFAULTS } from './constants';
-export { NodeCanvas } from '../canvas-config';
+export { NodeCanvas } from '../canvasConfig';
